Section_06/20_OTP Verification/pyproject.toml
[project]
name = "20-otp-verification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
]
